requirements.txt
torch>=2.0.0
torch-geometric
transformers>=4.28.0

pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
joblib>=1.2.0
psycopg2-binary  
imbalanced-learn>=0.10.0

matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

pyyaml
hashlib